changed project structure
seperated model code into its own folder. Broke apart sprawling utilities files into their own classes. Created a general utilities folder for backend utilities.
