+++ src/main/java/com/flippingutilities/model/OfferEvent.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.flippingutilities;
+package com.flippingutilities.model;
 
 
 import com.google.gson.annotations.SerializedName;
