+++ src/main/java/com/flippingutilities/ui/gehistorytab/GeHistoryTabOfferPanel.java
@@ -1,10 +1,9 @@
 package com.flippingutilities.ui.gehistorytab;
 
-import com.flippingutilities.OfferEvent;
-import com.flippingutilities.ui.utilities.UIUtilities;
+import com.flippingutilities.model.OfferEvent;
+
 import java.awt.BorderLayout;
 import java.awt.Color;
-import java.awt.FlowLayout;
 import java.util.List;
 import java.util.function.BiConsumer;
 import javax.swing.JCheckBox;
