+++ src/main/java/com/flippingutilities/ui/uiutilities/CustomColors.java
@@ -1,0 +1,11 @@
+package com.flippingutilities.ui.uiutilities;
+
+import java.awt.*;
+
+public class CustomColors {
+    public static final Color OUTDATED_COLOR = new Color(250, 74, 75);
+    public static final Color PROFIT_COLOR = new Color(255, 175, 55);
+    public static final Color DARK_GRAY_ALT_ROW_COLOR = new Color(35, 35, 35);
+    public static final Color DARK_GRAY = new Color(27, 27, 27);
+    public static final Color VIBRANT_YELLOW = new Color(230, 220, 6);
+}
