+++ src/test/java/com/flippingutilities/FlippingPluginTest.java
@@ -6,6 +6,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import com.flippingutilities.model.AccountData;
+import com.flippingutilities.model.OfferEvent;
 import net.runelite.api.GrandExchangeOfferState;
 import org.junit.Before;
 import org.junit.Test;
