+++ src/test/java/com/flippingutilities/HistoryManagerTest.java
@@ -32,6 +32,10 @@ import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+
+import com.flippingutilities.model.Flip;
+import com.flippingutilities.model.HistoryManager;
+import com.flippingutilities.model.OfferEvent;
 import net.runelite.api.GrandExchangeOfferState;
 import static org.junit.Assert.assertEquals;
 import org.junit.Before;
